HappyFace/requirements.txt
pygubu~=0.31
python-dotenv~=1.0.0
numpy~=1.24.3
opencv-python~=4.8.0.76
mediapipe~=0.10.3
keras~=2.13.1
tensorflow~=2.13.0
pandas~=2.1.0
mysql-connector-python
face-recognition
